+++ src/components/action/index.js
@@ -2,9 +2,7 @@ import React, { Component } from 'react';
 
 import Button from '../button';
 import Switch from '../switch';
-import DropdownColor from '../dropdown-color';
-import DropdownText from '../dropdown-text';
-//import DropdownText from '../dropdown-text';
+import { Dropdown, DropdownColor} from '../dropdown';
 import ActionTypes from 'butter-component-action-types';
 
 
@@ -22,7 +20,7 @@ export default class Action extends Component {
             (type === ActionTypes.BUTTON)?(<Button title={t(props.title)}/>):
             (type === ActionTypes.TEXT)?(<input type="text" value={value} onChange={this.apply}/>):
             (type === ActionTypes.PASSWORD)?(<input type="password"/>):
-            (type === ActionTypes.DROPDOWN)?(<DropdownText apply={this.apply} selected={value} {...props}/>):
+            (type === ActionTypes.DROPDOWN)?(<Dropdown apply={this.apply} selected={value} {...props}/>):
             (type === ActionTypes.COLOR)?(<DropdownColor apply={this.apply} selected={value} {...props}/>):
             (type === ActionTypes.SWITCH)?(<Switch apply={this.apply} selected={value}/>):
             (<b className="error">Couldn't find an apropiate action type {console.log('Could not find action', type, props)}</b>)
